+++ test/THSpec.hs
@@ -9,6 +9,7 @@
 {-# LANGUAGE UndecidableInstances #-}
 #if __GLASGOW_HASKELL__ >= 708
 {-# LANGUAGE EmptyCase #-}
+{-# LANGUAGE RoleAnnotations #-}
 #endif
 
 {-# OPTIONS_GHC -fno-warn-name-shadowing #-}
@@ -68,6 +69,9 @@ data IntFunD a b = IntFunD (IntFun a b)
 
 data Empty1 a b
 data Empty2 a b
+#if __GLASGOW_HASKELL__ >= 708
+type role Empty2 nominal nominal
+#endif
 
 -- Data families